feat: form become component
