+++ src/components/utils/greenLine.jsx
@@ -1,5 +1,9 @@
 function GreenLine() {
-  return <hr className="w-[70%] border-4 border-FI_green" />;
+  return (
+    <div className="w-full flex items-center justify-center">
+      <hr className="w-[70%] max-w-screen-lg border-2 border-FI_green" />
+    </div>
+  )
 }
 
 export default GreenLine;
